fix(types): resolve all 88 mypy errors across the codebase
- runoff: float() casts for numpy returns, assert for Optional params,
  isinstance checks for LandCover | str union
- visualization: type: ignore[assignment] for Figure | None from
  get_figure(), float() casts for numpy scalars, inline isinstance
  for proper type narrowing
- reports: assert for Optional arrays, type annotations for params
- morphometry: TYPE_CHECKING imports for forward references
- cli: fix list/set assignment, int() cast for return type

mypy now reports 0 errors (52 files checked).

+++ hydrolog/runoff/nash_iuh.py
@@ -862,11 +862,11 @@ class NashIUH:
             n_minus_1 = N_val - 1
             numerator = (n_minus_1**N_val) * np.exp(-n_minus_1)
             denominator = gamma(N_val)
-            return numerator / denominator
+            return float(numerator / denominator)
 
         def objective(N_val: float) -> float:
             """Objective function: f(N) - target = 0."""
-            return f_N_equation(N_val) - f_N_target
+            return float(f_N_equation(N_val) - f_N_target)
 
         # f(N) has maximum around N=2-3, then decreases
         # Typical range for f(N): 0.35 - 0.40
